requirements.txt
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
dominate==2.6.0
future==0.18.2
gast==0.3.3
grpcio==1.34.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
jsonpatch==1.27
jsonpointer==2.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
networkx==2.5
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==1.3.0
Pillow @ file:///tmp/build/80754af9/pillow_1603822255246/work
scikit-image==0.17.2
scipy==1.5.4
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
tifffile==2020.11.26
torch==1.5.0
torchvision==0.6.0
tqdm==4.54.1
urllib3==1.26.2
visdom==0.1.8.9
websocket-client==0.57.0
